Reportモデル、ReportViewモデル作成

+++ src/main/java/actions/EmployeeAction.java
@@ -44,6 +44,9 @@ public class EmployeeAction extends ActionBase {
 
     public void index() throws ServletException, IOException {
 
+        //管理者かどうかのチェック//追記
+        if (checkAdmin()) {
+
         //指定されたページ数の一覧表示するデータを取得
         int page = getPage();
         List<EmployeeView> employees = service.getPerPage(page);
@@ -68,7 +71,7 @@ public class EmployeeAction extends ActionBase {
         forward(ForwardConst.FW_EMP_INDEX);
 
     }
-
+}
     /**
      * 新規登録画面を表示する
      *
@@ -79,12 +82,16 @@ public class EmployeeAction extends ActionBase {
 
     public void entryNew() throws ServletException, IOException {
 
+        //管理者かどうかのチェック//  追記
+        if (checkAdmin()) {
+
         putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
         putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView()); //空の従業員インスタンス
 
         //新規登録画面を表示
         forward(ForwardConst.FW_EMP_NEW);
     }
+}
 
     /**
      * 新規登録を行う
@@ -97,7 +104,7 @@ public class EmployeeAction extends ActionBase {
     public void create() throws ServletException, IOException {
 
         //CSRF対策　tokenのチェック
-        if (checkToken()) {
+        if (checkAdmin() && checkToken()) {
 
             //パラメータの値を元に従業員情報のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
@@ -151,6 +158,10 @@ public class EmployeeAction extends ActionBase {
 
     public void show() throws ServletException, IOException {
 
+        //管理者かどうかのチェック
+        if (checkAdmin()) {
+
+        }
         //idを条件に従業員データを取得する
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
@@ -176,6 +187,9 @@ public class EmployeeAction extends ActionBase {
      */
     public void edit() throws ServletException, IOException {
 
+        //管理者かどうかのチェック
+        if (checkAdmin()) {
+
         //idを条件に従業員データを取得する
         EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
@@ -186,6 +200,7 @@ public class EmployeeAction extends ActionBase {
             return;
         }
 
+
         putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
         putRequestScope(AttributeConst.EMPLOYEE, ev); //取得した従業員情報
 
@@ -193,6 +208,7 @@ public class EmployeeAction extends ActionBase {
         forward(ForwardConst.FW_EMP_EDIT);
 
     }
+}
 
     /**
      * 更新を行う
@@ -202,7 +218,7 @@ public class EmployeeAction extends ActionBase {
     public void update() throws ServletException, IOException {
 
         //CSRF対策　tokenのチェック
-        if (checkToken()) {
+        if (checkAdmin() && checkToken()) {
             //パラメータの値を元に従業員情報のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
                     toNumber(getRequestParam(AttributeConst.EMP_ID)),
@@ -251,7 +267,7 @@ public class EmployeeAction extends ActionBase {
     public void destroy() throws ServletException, IOException {
 
         //CSRF対策　tokenのチェック
-        if (checkToken()) {
+        if (checkAdmin() && checkToken()) {
 
             //idを条件に従業員データを論理削除する
             service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
@@ -264,4 +280,28 @@ public class EmployeeAction extends ActionBase {
         }
     }
 
-}
+    /**
+     * ログイン中の従業員が管理者かどうかのチェックし、管理者でなければエラー画面を表示
+     * true: 管理者　false：　管理者ではない
+     */
+
+    private boolean checkAdmin() throws ServletException, IOException {
+
+        //セッションからログイン中の従業員情報を取得
+        EmployeeView ev = (EmployeeView) getSessionScope(AttributeConst.LOGIN_EMP);
+
+        //管理者でなければエラー画面を表示
+        if (ev.getAdminFlag() != AttributeConst.ROLE_ADMIN.getIntegerValue()) {
+
+            forward(ForwardConst.FW_ERR_UNKNOWN);
+            return false;
+
+        } else {
+            return true;
+
+
+        }
+    }
+
+    }
+
